Migrate utility functions and remove source modules
